Move all math to MathOperations.cs
